WheelDatalog
This program will (should) record how fast the robot can go based upon a
timer, encoder values, and the actual distance measured by red lines
every foam square. The robot will go the distance of the field, and the
datalog will consist of the set speed, both encoder values, current
distance, current battery voltage, and current time. These values are
recorded every tenth of a second. This should be useful when comparing
things like PID control and when exactly we should switch out the
battery. Also, there is no waitForStart() so you don't need a computer
on hand to run it (but reading it might be tricky, I'm not sure). The
datalog can be copied from the RobotC IDE into an excel spreadsheet.